fastMCP-POC/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
cyclopts==3.22.2
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.16
docutils==0.21.2
ecdsa==0.19.1
email_validator==2.2.0
eunomia-ai==0.3.9
eunomia-core==0.3.9
eunomia-mcp==0.3.9
eunomia-sdk==0.3.9
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastmcp @ git+https://github.com/jlowin/fastmcp.git@b4cd2857bb8341424a4f8ac2f8aec595e7c34703
filetype==1.2.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
langchain-core==0.3.69
langchain-google-genai==2.1.8
langsmith==0.4.6
lazy-object-proxy==1.12.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.13.1
mdurl==0.1.2
more-itertools==10.7.0
oauthlib==3.3.1
openai==1.102.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
orjson==3.11.0
packaging==25.0
parse==1.20.2
pathable==0.4.4
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pyperclip==1.9.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rich-rst==1.3.1
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.26.0
rsa==4.9.1
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sse-starlette==2.4.1
starlette==0.47.1
tenacity==9.1.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uv==0.8.13
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.1
zstandard==0.23.0
